requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
WTForms==3.1.2
Flask-SQLAlchemy==3.1.1
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
Werkzeug==2.3.7